Se agregan endpoints de prueba para estudiantes, asignaturas, demanda, preins.
Se agregan actuators de spring boot

se agrega la estructura de configuración optaplanner

+++ PreinscripcionAuto/src/main/java/co/edu/udistrital/sga/preinscripcion/auto/persistence/entities/oracle/Acest.java
@@ -84,7 +84,7 @@ public class Acest implements Serializable {
 	private BigDecimal estPbm;
 
 	@Column(name="EST_PEN_NRO")
-	private BigDecimal estPenNro;
+	private Integer estPenNro;
 
 	@Column(name="EST_PORCENTAJE")
 	private BigDecimal estPorcentaje;
